+++ loglan_core/addons/word_selector.py
@@ -15,7 +15,7 @@ from .base_selector import BaseSelector
 from .definition_selector import DefinitionSelector
 from .filters import filter_word_by_event_id
 from ..key import BaseKey
-from ..service.connect_tables import t_connect_words
+from loglan_core.relationships import t_connect_words
 from ..type import BaseType
 from ..word import BaseWord
