+++ loglan_core/base.py
