+++ loglan_core/event.py
@@ -12,7 +12,7 @@ from sqlalchemy.orm import Mapped, mapped_column
 from sqlalchemy.orm import relationship
 
 from .service.annotated_types import str_064, str_016
-from .service.base import BaseModel
+from .base import BaseModel
 from .service.table_names import T_NAME_EVENTS
 
 if TYPE_CHECKING:
