+++ loglan_core/key.py
@@ -11,8 +11,8 @@ from sqlalchemy.orm import mapped_column, Mapped
 from sqlalchemy.orm import relationship
 
 from .service.annotated_types import str_016, str_064
-from .service.base import BaseModel
-from .service.connect_tables import t_connect_keys
+from .base import BaseModel
+from .relationships import t_connect_keys
 from .service.table_names import T_NAME_KEYS
 
 if TYPE_CHECKING:
