+++ loglan_core/relationships.py
@@ -20,7 +20,7 @@ and includes indices on key columns to enhance query performance.
 
 from sqlalchemy import Column, ForeignKey, Table, Index
 
-from .table_names import (
+from loglan_core.service.table_names import (
     T_NAME_AUTHORS,
     T_NAME_KEYS,
     T_NAME_WORDS,
@@ -29,7 +29,7 @@ from .table_names import (
     T_NAME_CONNECT_WORDS,
     T_NAME_CONNECT_KEYS,
 )
-from ..service.base import BaseModel as Base
+from .base import BaseModel as Base
 
 t_connect_authors = Table(
     T_NAME_CONNECT_AUTHORS,
