+++ loglan_core/syllable.py
@@ -5,7 +5,7 @@ This module contains a basic Syllable Model
 from sqlalchemy.orm import mapped_column, Mapped
 
 from .service.annotated_types import str_008, str_032
-from .service.base import BaseModel
+from .base import BaseModel
 from .service.table_names import T_NAME_SYLLABLES
