+++ loglan_core/type.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING
 from sqlalchemy.orm import relationship, mapped_column, Mapped
 
 from .service.annotated_types import str_016, str_255
-from .service.base import BaseModel
+from .base import BaseModel
 from .service.table_names import T_NAME_TYPES
 
 if TYPE_CHECKING:
